backend/pyproject.toml
[project]
name = "zhiyan-backend"
version = "0.1.0"
description = "知演 ZhiYan — AI PPT 生成智能体后端"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.8",
    "markitdown[all]>=0.1.2",
    "pydantic>=2.12.5",
    "pydantic-ai[anthropic,google,openai]>=1.58.0",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.22",
    "python-pptx>=1.0.2",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.0",
]
